test: Added TestConfigLoading roundtrip tests for correlation stop/TP c…
- "agents/signals/tests/test_correlation.py"
- "agents/signals/strategies/correlation.py"

GSD-Task: S02/T02

+++ agents/signals/strategies/correlation.py
@@ -371,10 +371,10 @@ class CorrelationStrategy(SignalStrategy):
         # If opposite: divergence (sign indicates bullish/bearish)
         if price_pct > 0 and oi_pct < 0:
             # Price up, OI down -> bearish divergence (distribution)
-            return -(abs(price_pct) + abs(oi_pct)) / 2
+            return float(-(abs(price_pct) + abs(oi_pct)) / 2)
         elif price_pct < 0 and oi_pct > 0:
             # Price down, OI up -> bullish divergence (accumulation)
-            return (abs(price_pct) + abs(oi_pct)) / 2
+            return float((abs(price_pct) + abs(oi_pct)) / 2)
         else:
             return 0.0
